requirements.txt
Django==3.1.7
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-location-field==2.1.0
django-safedelete==0.5.6
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.25.1


